fixing build for gcc 4.6.0, x64 (Fedora)
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1118 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/util/clock.h
@@ -64,7 +64,7 @@ public:
 	explicit Clock(TTime iStartTime = 0);
 	void reset(TTime iStartTime = 0);
 
-	TTime time() const { return startTime_ + resolution_ * this->tick(); }
+	TTime time() const { return startTime_ + resolution_ * static_cast<double>(this->tick()); }
 	TTick tick() const { return impl::ClockImpl::tick() - startTick_; }
 
 	TTick frequency() const { return frequency_; }
